auth and participants

+++ src/LeftTuesday/Services/ConceptService.cs
@@ -10,34 +10,53 @@ namespace LeftTuesday.Services
     public class ConceptService
     {
         private ConceptRepository _conceptRepo;
+        private IsOwnerService _isOwnerService;
 
-        public ConceptService(ConceptRepository concpetRepo)
+        public ConceptService(ConceptRepository concpetRepo,
+            IsOwnerService isOwnerService)
         {
             _conceptRepo = concpetRepo;
+            _isOwnerService = isOwnerService;
         }
 
-        public (Exception, long) CreateConcept(Concept concept)
-        {
-            //Todo add validation
+        public (Exception, long) CreateConcept(Concept concept, long ownerId)
+        {     
+            var(error, conceptId) = _conceptRepo.AddConcpet(concept);
+            if(error != null)
+            {
+                return (error, default);
+            }
 
-            return _conceptRepo.AddConcpet(concept);
+            return _conceptRepo.AddOwner(conceptId, ownerId);
         }
 
-        public (Exception, Concept) GetConcept(int conceptId)
+        public (Exception, Concept) GetConcept(int conceptId, long ownerId)
         {
-            //Todo add validation
+            if(!_isOwnerService.IsConceptOwner(conceptId, ownerId))
+            {
+                return (new Exception("Verification failed"), null);
+            }
+
             return _conceptRepo.GetConcpet(conceptId);
         }
 
         public (Exception, List<Concept>) GetConcepts()
         {
-            //Todo add validation
             return _conceptRepo.GetConcpets();
         }
 
-        public (Exception, Concept) UpdateConcept(Concept concept)
+        public (Exception, List<ConceptOwner>) GetAllConceptOwners()
         {
-            //Todo add validation
+            return _conceptRepo.GetAllConcpetOwners();
+        }
+
+        public (Exception, Concept) UpdateConcept(Concept concept, long ownerId)
+        {
+            if (!_isOwnerService.IsConceptOwner(concept.Id, ownerId))
+            {
+                return (new Exception("Verification failed"), null);
+            }
+
             var (error, original) = _conceptRepo.GetConcpet(concept.Id);
 
             if(error!= null)
@@ -63,8 +82,12 @@ namespace LeftTuesday.Services
             return _conceptRepo.UpdateConcpet(original);
         }
 
-        public (Exception, bool) DeleteConcept(int conceptId)
+        public (Exception, bool) DeleteConcept(int conceptId, long ownerId)
         {
+            if (!_isOwnerService.IsConceptOwner(conceptId, ownerId))
+            {
+                return (new Exception("Verification failed"), false);
+            }
             //Todo cascade delete?
             return _conceptRepo.DeleteConcpet(conceptId);
         }
